updates after rehearsal
Q: Is anyone checking for Pull Requests? It says at the beginning that SVs can submit questions that way
Q: In the “SV Shift Plan” section, it says that the Pacific time band shift starts 2 hours before the Pacific ICSE time band (i.e. the shift starts at 2200, the ICSE session starts at 0000), but the others start 1 hour before. Is this a mistake?
Added slack channels and descriptions - check if they’re right
#### Managing 24-hour help desk - this is where we need to point SVs to resources for the problems they find
#### Helping Live Casting - removed a bunch of stuff we learned during the rehearsals that SVs will not be doing - what about checking for free riders - do SVs need to do that and how do they do that?
I did not always know how to preserve formatting, so my changes might need to be cleaned up a bit to look nice.
Q: Will there be green room Slack channels in the ICSE2020 workspace next week and will SVs be invited to them? If so, add to example scenario.